refactor of indexing collection
