+++ src/main/java/io/anserini/ltr/WebCollectionLtrDataGenerator.java
@@ -1,6 +1,6 @@
 package io.anserini.ltr;
 
-import io.anserini.index.IndexWebCollection;
+import io.anserini.index.IndexThreads;
 import io.anserini.ltr.feature.FeatureExtractors;
 import io.anserini.rerank.Reranker;
 import io.anserini.rerank.RerankerContext;
@@ -44,9 +44,9 @@ public class WebCollectionLtrDataGenerator implements Reranker{
     LOG.info("Beginning rerank");
     for (int i =0; i < docs.documents.length; i++ ) {
       try {
-        Terms terms = reader.getTermVector(docs.ids[i], IndexWebCollection.FIELD_BODY);
+        Terms terms = reader.getTermVector(docs.ids[i], IndexThreads.FIELD_BODY);
         float[] features = this.extractorChain.extractAll(documents[i], terms, context);
-        String docId = documents[i].get(IndexWebCollection.FIELD_ID);
+        String docId = documents[i].get(IndexThreads.FIELD_ID);
         // QREL 0 in this case, will be assigned if needed later
         //qid
         BaseFeatureExtractor.writeFeatureVector(out,qid, this.qrels.getRelevanceGrade(qid, docId), docId,  features);
